+++ config/site/site.config.js
@@ -1,13 +1,44 @@
 const startCase = require('lodash.startcase');
 const fs = require('fs');
 const path = require('path');
-const navConfig = require('./navigation.config');
 
 const npmPackage = JSON.parse(fs.readFileSync(path.join(process.cwd(), 'package.json')));
 
 const siteConfig = {
   /* The navigation configuration. */
-  navConfig,
+  navConfig: {
+    navigation: {
+      /* The first page to route to for the site. */
+      index: '/home',
+      /** List of top level nav items.
+       *   Link config options:
+       *      path: Path to the link.
+       *      text: The text to display on the navigation link.
+       *      pageTypes: The page extension(s) that should be displayed under this link.
+       *      capabilities: an object describing the capabilities of all pages listed under the specified path.
+       *          devTools: display development tools to allow switching between locales and themes.
+       */
+      links: [{
+        path: '/home',
+        text: 'Home',
+        pageTypes: ['home'],
+      }, {
+        path: '/components',
+        text: 'Components',
+        pageTypes: ['doc'],
+        capabilities: {
+          devTools: true,
+        },
+      }, {
+        path: '/tests',
+        text: 'Tests',
+        pageTypes: ['test'],
+        capabilities: {
+          devTools: true,
+        },
+      }],
+    },
+  },
 
   /** These options are used to find the pages to serve via terra-dev-site.
    * The file extensions pulled in are 'md' extensions and any extension defined in the resolve extensions set in the webpack config.
@@ -50,31 +81,16 @@ const siteConfig = {
     /* Path to the favicon for the site. */
     favicon: path.join(__dirname, '..', '..', 'terra-favicon', '32px', 'favicon.ico'),
 
-    /** The themes to supply the ThemeProvider which allows the site to switch
-      * between themes. Providing multiple enables the theme utility to display
-      * in the header's toolbar.
-      */
-    themes: {
-      'Default Theme': '',
-    },
-
     /** The default theme of the site. Note, this value should be a key that was
-     * supplied to the themes object. The open-sourced theme is the default theme.
+     * supplied to the themes object.
      */
-    defaultTheme: 'Default Theme',
-
-    /** The locales to supply Base with, which allows the site to switch
-     * between locales. Defaulted to the supported locals list in terra-18n.
-     *
-     * NOTE: This configuration has been deprecated as of terra-dev-site 6.23.0
-     * locales: i18nSupportedLocales,
-     */
+    // defaultTheme: 'terra-default-theme',
 
     /* The default locale of the site. 'en' is the default theme. */
-    defaultLocale: 'en',
+    // defaultLocale: 'en',
 
     /* The default direction of the site. 'ltr' is the default direction. */
-    defaultDirection: 'ltr',
+    // defaultDirection: 'ltr',
 
     /** This section allows you to set custom extensions.
      * Extensions will be launched in a modal manager. All fields are required.