+++ dev-site-config/site.config.js
@@ -1,24 +1,60 @@
 const fs = require('fs');
-const navConfig = require('./navigation.config');
 
 const html = fs.readFileSync(require.resolve('./head.html'), 'utf8');
 
 const siteConfig = {
-  navConfig,
-  sideEffectImports: ['./terra-dev-site-test-theme.scss'],
+  navConfig: {
+    navigation: {
+      index: '/home',
+      links: [{
+        path: '/home',
+        text: 'Home',
+        pageTypes: ['home'],
+      }, {
+        path: '/dev_tools',
+        text: 'Developer Tools',
+        pageTypes: ['tool'],
+      }, {
+        path: '/single-page-test',
+        text: 'Single Page Test',
+        pageTypes: ['spt'],
+        capabilities: {
+          devTools: true,
+        },
+      }, {
+        path: '/secondary-nav-test',
+        text: 'Secondary Nav Test',
+        pageTypes: ['snt'],
+        capabilities: {
+          devTools: true,
+        },
+      }, {
+        path: '/folder-first',
+        text: 'Folder First Test',
+        pageTypes: ['ff'],
+        capabilities: {
+          devTools: true,
+        },
+      }, {
+        path: '/empty',
+        text: 'Empty',
+        pageTypes: ['empty'],
+      }, {
+        path: '/test',
+        text: 'Test',
+        pageTypes: ['test'],
+        capabilities: {
+          devTools: true,
+        },
+      }],
+    },
+  },
 
   appConfig: {
     headHtml: [
       '<script> console.log("Inline head html script") </script>',
       html,
     ],
-
-    themes: {
-      'Default Theme': '',
-      'Clinical-lowlight-theme': 'clinical-lowlight-theme',
-      'Terra Dev Site Test Theme': 'terra-dev-site-test-theme',
-      'Orion Fusion Theme': 'orion-fusion-theme',
-    },
 
     extensions: [
       {